ref(symcache): Remove old format
This removes the old symcache format and performs
a slew of refactorings that follow from that.
