+++ symbolic-symcache/src/error.rs
@@ -5,7 +5,7 @@ use thiserror::Error;
 /// After a SymCache was successfully parsed via [`SymCache::parse`](crate::new::SymCache::parse), an Error that occurs during
 /// access of any data indicates either corruption of the serialized file, or a bug in the
 /// converter/serializer.
-#[derive(Debug, Error, Clone, Copy)]
+#[derive(Debug, Error, Clone, Copy, PartialEq, Eq)]
 #[non_exhaustive]
 pub enum ErrorKind {
     /// The buffer is not correctly aligned.
@@ -41,6 +41,16 @@ pub struct Error {
 }
 
 impl Error {
+    /// Creates a new SymCache error from a known kind of error as well as an
+    /// arbitrary error payload.
+    pub(crate) fn new<E>(kind: ErrorKind, source: E) -> Self
+    where
+        E: Into<Box<dyn std::error::Error + Send + Sync>>,
+    {
+        let source = Some(source.into());
+        Self { kind, source }
+    }
+
     /// Returns the corresponding [`SymCacheErrorKind`] for this error.
     pub fn kind(&self) -> ErrorKind {
         self.kind
