--- symbolic-symcache/src/old/mod.rs
@@ -1,9 +1,0 @@
-//! Provides support to read the old SymCache format.
-
-mod cache;
-mod error;
-
-pub mod format;
-
-pub use cache::*;
-pub use error::*;
