+++ symbolic-symcache/src/raw.rs
