+++ symbolic-symcache/src/transform/bcsymbolmap.rs
