+++ symbolic-symcache/src/transform/il2cpp.rs
