+++ symbolic-symcache/src/writer.rs
@@ -16,7 +16,7 @@ use symbolic_common::Language;
 use symbolic_il2cpp::usym::{UsymSourceRecord, UsymSymbols};
 
 use super::{raw, transform};
-use crate::{SymCacheError, SymCacheErrorKind};
+use crate::{Error, ErrorKind};
 
 /// The SymCache Converter.
 ///
@@ -117,21 +117,20 @@ impl SymCacheConverter {
 
     /// This processes the given [`ObjectLike`] object, collecting all its functions and line
     /// information into the converter.
-    pub fn process_object<'d, 'o, O>(&mut self, object: &'o O) -> Result<(), SymCacheError>
+    pub fn process_object<'d, 'o, O>(&mut self, object: &'o O) -> Result<(), Error>
     where
         O: ObjectLike<'d, 'o>,
         O::Error: std::error::Error + Send + Sync + 'static,
     {
         let session = object
             .debug_session()
-            .map_err(|e| SymCacheError::new(SymCacheErrorKind::BadDebugFile, e))?;
+            .map_err(|e| Error::new(ErrorKind::BadDebugFile, e))?;
 
         self.set_arch(object.arch());
         self.set_debug_id(object.debug_id());
 
         for function in session.functions() {
-            let function =
-                function.map_err(|e| SymCacheError::new(SymCacheErrorKind::BadDebugFile, e))?;
+            let function = function.map_err(|e| Error::new(ErrorKind::BadDebugFile, e))?;
 
             self.process_symbolic_function(&function);
         }
@@ -322,12 +321,12 @@ impl SymCacheConverter {
 
     #[cfg(feature = "il2cpp")]
     /// Processes a set of [`UsymSymbols`], passing all mapped symbols into the converter.
-    pub fn process_usym(&mut self, usym: &UsymSymbols) -> Result<(), SymCacheError> {
+    pub fn process_usym(&mut self, usym: &UsymSymbols) -> Result<(), Error> {
         // Assume records they are sorted by address; There's a test that guarantees this
 
         let debug_id = usym
             .id()
-            .map_err(|e| SymCacheError::new(SymCacheErrorKind::BadFileHeader, e))?;
+            .map_err(|e| Error::new(ErrorKind::HeaderTooSmall, e))?;
         self.set_debug_id(debug_id);
 
         let arch = usym.arch().unwrap_or_default();
