+++ symbolic-symcache/tests/test_cache.rs
@@ -10,15 +10,17 @@ fn test_load_header_linux() -> Result<(), Error> {
     let symcache = SymCache::parse(&buffer)?;
     insta::assert_debug_snapshot!(symcache, @r###"
     SymCache {
-        version: 2,
+        version: 7,
         debug_id: DebugId {
             uuid: "c0bcc3f1-9827-fe65-3058-404b2831d9e6",
             appendix: 0,
         },
         arch: Amd64,
-        has_line_info: true,
-        has_file_info: true,
-        functions: 1955,
+        files: 55,
+        functions: 697,
+        source_locations: 8236,
+        ranges: 6762,
+        string_bytes: 52180,
     }
     "###);
     Ok(())
@@ -38,15 +40,17 @@ fn test_load_header_macos() -> Result<(), Error> {
     let symcache = SymCache::parse(&buffer)?;
     insta::assert_debug_snapshot!(symcache, @r###"
     SymCache {
-        version: 2,
+        version: 7,
         debug_id: DebugId {
             uuid: "67e9247c-814e-392b-a027-dbde6748fcbf",
             appendix: 0,
         },
         arch: Amd64,
-        has_line_info: true,
-        has_file_info: true,
-        functions: 1863,
+        files: 36,
+        functions: 639,
+        source_locations: 6033,
+        ranges: 4591,
+        string_bytes: 42829,
     }
     "###);
     Ok(())
@@ -64,7 +68,7 @@ fn test_load_functions_macos() -> Result<(), Error> {
 fn test_lookup() -> Result<(), Error> {
     let buffer = ByteView::open(fixture("symcache/current/macos.symc"))?;
     let symcache = SymCache::parse(&buffer)?;
-    let line_infos = symcache.lookup(4_458_187_797 - 4_458_131_456)?;
+    let line_infos = symcache.lookup(4_458_187_797 - 4_458_131_456);
     insta::assert_debug_snapshot!("lookup", &line_infos);
 
     Ok(())
